add multi-path access in LocalFileRepository & add async model

+++ file-repository-sample/src/main/java/org/coodex/filerepository/sample/CtrCryptoAccessSample.java
@@ -5,36 +5,68 @@ import org.coodex.filerepository.api.FileMetaInf;
 import org.coodex.filerepository.api.IFileRepository;
 import org.coodex.filerepository.api.StoredFileMetaInf;
 import org.coodex.filerepository.ext.callback.CtrCryptoReadCallback;
-import org.coodex.filerepository.ext.callback.CtrCryptoWriteCallback;
 import org.coodex.filerepository.local.HashPathGenerator;
 import org.coodex.filerepository.local.LocalFileRepository;
+import org.coodex.filerepository.local.LocalRepositoryPath;
 import org.coodex.util.Profile;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.*;
-import java.security.InvalidAlgorithmParameterException;
-import java.security.InvalidKeyException;
+import java.util.ArrayList;
 import java.util.Base64;
+import java.util.List;
+import java.util.Scanner;
 
 public class CtrCryptoAccessSample {
     private static Logger log = LoggerFactory.getLogger(CtrCryptoAccessSample.class);
 
     private static Profile config = Profile.get("config.properties");
 
-    public static void main(String[] args) throws IOException, InvalidAlgorithmParameterException, InvalidKeyException {
+    public static void main(String[] args) throws Throwable {
+        String[] paths = config.getStrList("file.repository.path.list");
+        List<LocalRepositoryPath> pathList = new ArrayList<>();
+        for (String pathName : paths) {
+            String bathPath = config.getString("file.repository.path." + pathName);
+            pathList.add(new LocalRepositoryPath(
+                    config.getString("file.repository.path." + pathName),
+                    config.getBool("file.repository.path." + pathName + ".read", true),
+                    config.getBool("file.repository.path." + pathName + ".write", true)
+            ));
+        }
         IFileRepository fileRepository = new LocalFileRepository(
-                config.getString("file.repository.path", "/Users/sujiwu/Downloads/temp/"),
+                pathList.toArray(new LocalRepositoryPath[0]),
                 new HashPathGenerator());
         String fileId = saveFile(fileRepository);
-        log.debug("file id: {}", fileId);
+        log.info("file saved, id: {}", fileId);
+        Scanner input = new Scanner(System.in);
+        log.info("continue ? (y/n)");
+        String i = input.next();
+        if (!i.toLowerCase().equals("y")) {
+            return;
+        }
+
         StoredFileMetaInf fileMetaInf = fileRepository.getMetaInf(fileId);
-        log.debug("file meta-inf: {}", JSON.toJSONString(fileMetaInf));
+        log.info("file meta-inf: {}", JSON.toJSONString(fileMetaInf));
+        log.info("continue ? (y/n)");
+        i = input.next();
+        if (!i.toLowerCase().equals("y")) {
+            return;
+        }
+
         getFile(fileId, fileMetaInf.getFileName() + "." + fileMetaInf.getExtName(), fileRepository);
+        log.info("get file, id: {}", fileId);
+        log.info("continue ? (y/n)");
+        i = input.next();
+        if (!i.toLowerCase().equals("y")) {
+            return;
+        }
+
         deleteFile(fileId, fileRepository);
+        log.info("store file deleted, fileId: {}", fileId);
     }
 
-    private static String saveFile(IFileRepository fileRepository) throws IOException {
+    private static String saveFile(IFileRepository fileRepository) throws Throwable {
         File file = new File(config.getString("sample.save.file"));
         FileMetaInf fileMetaInf = new FileMetaInf();
         fileMetaInf.setClientId(CtrCryptoAccessSample.class.getSimpleName());
@@ -44,19 +76,30 @@ public class CtrCryptoAccessSample {
         fileMetaInf.setFileSize(file.length());
         InputStream inputStream = new FileInputStream(file);
         try {
-            return fileRepository.save(fileMetaInf,
-                    new CtrCryptoWriteCallback(inputStream, Base64.getDecoder().decode(config.getString("aes.ctr.key"))));
+//            return fileRepository.save(fileMetaInf,
+//                    new CtrCryptoWriteCallback(inputStream, Base64.getDecoder().decode(config.getString("aes.ctr.key"))));
+            return fileRepository.asyncSave(inputStream, fileMetaInf, ((success, fileId, t) -> {
+                try {
+                    inputStream.close();
+                    if (!success) {
+                        log.error(t.getLocalizedMessage(), t);
+                    } else {
+                        log.debug("save file {} in async model", fileId);
+                    }
+                } catch (IOException e) {
+                    log.error(e.getLocalizedMessage(), e);
+                }
+            }));
         } finally {
-            inputStream.close();
+//            inputStream.close();
         }
     }
 
-    private static StoredFileMetaInf getFileMetaInf(String fileId, IFileRepository fileRepository) {
+    private static StoredFileMetaInf getFileMetaInf(String fileId, IFileRepository fileRepository) throws Throwable {
         return fileRepository.getMetaInf(fileId);
     }
 
-    private static void getFile(String fileId, String fileName, IFileRepository fileRepository) throws IOException,
-            InvalidAlgorithmParameterException, InvalidKeyException {
+    private static void getFile(String fileId, String fileName, IFileRepository fileRepository) throws Throwable {
         File file = new File(config.getString("sample.get.file.path") + "/" + fileName);
         File filePath = file.getParentFile();
         if (!filePath.exists()) {
@@ -68,7 +111,14 @@ public class CtrCryptoAccessSample {
                 new CtrCryptoReadCallback(outputStream, Base64.getDecoder().decode(config.getString("aes.ctr.key"))));
     }
 
-    private static void deleteFile(String fileId, IFileRepository fileRepository) {
-        fileRepository.delete(fileId);
+    private static void deleteFile(String fileId, IFileRepository fileRepository) throws Throwable {
+//        fileRepository.delete(fileId);
+        fileRepository.asyncDelete(fileId, ((success, fileId1, t) -> {
+            if (!success) {
+                log.error(t.getLocalizedMessage(), t);
+            } else {
+                log.debug("delete file {} in async model", fileId1);
+            }
+        }));
     }
 }
